100 - Move OTelEvent into otel_to_pv_types.py
